Solve the mystery of the small index adjustments in NodeListManager
There were two reasons for small index adjustments: the first is that
for initial renders, the firstIndex that we receive really is the index
that we want to start adding at, since nothing exists. When we're
adding onto an existing list, we want to use the next index instead.

The second adjustment was made for going in the reverse direction. The
correction was automatically happening due to the math being done
that converted the 0-indexed value of `n` into a 1-indexed value by
using the `addDelta`